pyproject.toml
[project]
name = "2026-03030-vml-3d-renderer"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "google-generativeai>=0.7.0",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.19",
]

[project.scripts]
renderer = "renderer.server:start"

[tool.poetry]
packages = [{include = "renderer", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
